refactor(frontend): move swr hooks into application layer
