+++ frontend/src/modules/chats/application/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./chats.hooks";
