+++ frontend/src/modules/documents/application/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./documents.hooks";
