+++ frontend/src/modules/health/application/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./health.hooks";
