+++ frontend/src/modules/health/index.ts
@@ -1,1 +1,2 @@
+export * from "./application";
 export * from "./infrastructure";
