+++ frontend/src/modules/languages/application/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./languages.hooks";
