+++ frontend/src/modules/languages/index.ts
@@ -1,1 +1,2 @@
+export * from "./application";
 export * from "./domain";
