+++ frontend/src/modules/perspectives/application/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./perspectives.hooks";
