+++ frontend/src/modules/perspectives/index.ts
@@ -1,2 +1,3 @@
+export * from "./application";
 export * from "./domain";
 export * from "./infrastructure";
