+++ frontend/src/modules/tags/application/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./tags.hooks";
