+++ frontend/src/modules/tags/index.ts
@@ -1,2 +1,3 @@
+export * from "./application";
 export * from "./domain";
 export * from "./infrastructure";
