+++ frontend/src/modules/users/application/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./users.hooks";
