+++ frontend/src/modules/users/application/users.hooks.ts
@@ -1,0 +1,19 @@
+import useSWR from "swr";
+
+import { USERS_ENDPOINT } from "@/shared/config";
+import { swrJsonFetcher } from "@/shared/infrastructure/http";
+
+export function useUsersStatistics() {
+   const { data, error, isLoading, isValidating, mutate } = useSWR(
+      `${USERS_ENDPOINT}/statistics`,
+      swrJsonFetcher
+   );
+
+   return {
+      statistics: data?.data,
+      error,
+      isLoading,
+      isValidating,
+      mutate
+   };
+}
